Issue242_1: WeblogCustomTheme and WeblogShareTheme merged, stylesheet edits now being done on template edits page
