--- app/src/main/java/org/apache/roller/weblogger/business/themes/WeblogCustomTheme.java
@@ -1,129 +1,0 @@
-/*
- * Licensed to the Apache Software Foundation (ASF) under one or more
- *  contributor license agreements.  The ASF licenses this file to You
- * under the Apache License, Version 2.0 (the "License"); you may not
- * use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.  For additional information regarding
- * copyright in this work, please see the NOTICE file in the top level
- * directory of this distribution.
- *
- * Source file modified from the original ASF source; all changes made
- * are also under Apache License.
- */
-
-package org.apache.roller.weblogger.business.themes;
-
-import org.apache.roller.weblogger.WebloggerException;
-import org.apache.roller.weblogger.business.WeblogManager;
-
-import java.util.Date;
-import java.util.List;
-import org.apache.roller.weblogger.pojos.Theme;
-import org.apache.roller.weblogger.pojos.ThemeTemplate;
-import org.apache.roller.weblogger.pojos.ThemeTemplate.ComponentType;
-import org.apache.roller.weblogger.pojos.Weblog;
-import org.apache.roller.weblogger.pojos.WeblogTheme;
-
-
-/**
- * A WeblogTheme custom defined by the weblog owner.
- */
-public class WeblogCustomTheme extends WeblogTheme {
-
-    public WeblogCustomTheme(WeblogManager manager, Weblog weblog) {
-        super(manager, weblog);
-    }
-
-    public String getId() {
-        return CUSTOM;
-    }
-    
-    public String getName() {
-        return CUSTOM;
-    }
-
-    public String getType() {
-        return CUSTOM;
-    }
-
-    public String getDescription() {
-        return CUSTOM;
-    }
-
-    public String getAuthor() {
-        return "N/A";
-    }
-    
-    public Date getLastModified() {
-        return this.weblog.getLastModified();
-    }
-    
-    public boolean isEnabled() {
-        return true;
-    }
-
-    public int compareTo(Theme other) {
-        return getName().compareTo(other.getName());
-    }
-
-    /**
-     * Get the collection of all templates associated with this Theme.
-     */
-    public List<? extends ThemeTemplate> getTemplates() throws WebloggerException {
-        return weblogManager.getTemplates(this.weblog);
-    }
-    
-
-    /**
-     * Lookup the default template.
-     * Returns null if the template cannot be found.
-     */
-    public ThemeTemplate getDefaultTemplate() throws WebloggerException {
-        return weblogManager.getTemplateByAction(this.weblog, ComponentType.WEBLOG);
-    }
-    
-    
-    /**
-     * Lookup the specified template by action.
-     * Returns null if the template cannot be found.
-     */
-    public ThemeTemplate getTemplateByAction(ComponentType action) throws WebloggerException {
-        if (action == null) {
-            return null;
-        }
-        return weblogManager.getTemplateByAction(this.weblog, action);
-    }
-    
-    
-    /**
-     * Lookup the specified template by name.
-     * Returns null if the template cannot be found.
-     */
-    public ThemeTemplate getTemplateByName(String name) throws WebloggerException {
-        if (name == null) {
-            return null;
-        }
-        return weblogManager.getTemplateByName(this.weblog, name);
-    }
-    
-    
-    /**
-     * Lookup the specified template by link.
-     * Returns null if the template cannot be found.
-     */
-    public ThemeTemplate getTemplateByLink(String link) throws WebloggerException {
-        if (link == null) {
-            return null;
-        }
-        return weblogManager.getTemplateByLink(this.weblog, link);
-    }
-
-}
