+++ app/src/main/resources/ApplicationResources.properties
@@ -1023,32 +1023,14 @@ statCount.weblogDayHits=Today''s hit count
 
 # ------------------------------------------------------------------ Stylesheet Editor
 
-stylesheetEdit.title=Stylesheet
-stylesheetEdit.subtitle=Edit weblog custom stylesheet
-stylesheetEdit.tip=This form allows you to edit the stylesheet for your theme.
-stylesheetEdit.revertTip=If you run into any problems you don''t know how to \
-fix you can start over at any time by clicking the restore button.  Please <b>backup</b> your stylesheet!
-stylesheetEdit.revertTip1=<font color="red">If you have not customized your stylesheet here, \
-please use the <em>Shared Theme</em> default by clicking the delete stylesheet button. \
-This will ensure you are using the latest version and any changes will be picked up.</font> 
+templateEdit.revert=Restore Template
+templateEdit.delete=Delete Template
 
-stylesheetEdit.revert=Restore Stylesheet
-stylesheetEdit.delete=Delete Stylesheet
+templateEdit.revert.success=Template reverted to shared default successfully.
+templateEdit.default.success=Template override deleted successfully.  Your theme is now using the Shared Theme default.
 
-stylesheetEdit.create.success=Custom stylesheet created successfully.
-stylesheetEdit.save.success=Stylesheet updated successfully.
-stylesheetEdit.revert.success=Stylesheet reverted successfully.
-stylesheetEdit.default.success=Stylesheet deleted successfully.  Your theme is now using the Shared Theme default.
-
-stylesheetEdit.noOverridableStylesheetAvailable=\
-The theme you''re using does not define a customizable stylesheet in its descriptor.  You can choose another shared theme, or for \
-  custom themes, create the stylesheet on the templates tab instead.
-
-stylesheetEdit.canRecreateStylesheetOverride=Note: clicking the stylesheet tab again will \
-recreate a custom override stylesheet from the theme default.
-
-stylesheetEdit.confirmDelete=Are you sure you want to delete your stylesheet?
-stylesheetEdit.confirmRevert=Are you sure you want to revert your stylesheet to the theme default?
+templateEdit.confirmDelete=Are you sure you want to delete your stylesheet?
+templateEdit.confirmRevert=Are you sure you want to revert your stylesheet to the theme default?
 
 # ------------------------------------------------------------------ Tabbed Menu
 
@@ -1059,7 +1041,6 @@ tabbedmenu.weblog.categories=Categories
 tabbedmenu.bookmarks.allFolders=Blogroll
 
 tabbedmenu.design=Design
-tabbedmenu.design.stylesheet=Stylesheet
 
 tabbedmenu.website=Preferences
 tabbedmenu.website.settings=Settings
