+++ app/src/main/resources/ApplicationResources_de.properties
@@ -583,13 +583,6 @@ searchSummary.text=Angezeigt: {0} - {1} von insgesamt {2} Suchergebnissen
 statCount.weblogCommentCountType=Weblog Kommentar Anzahl
 statCount.weblogDayHits=Heutiger Hit Z\u00E4hler
 statCount.weblogEntryCommentCountType=Weblog Eintrag Kommentar Anzahl
-stylesheetEdit.revert=Stylesheet zur\u00FCcksetzen
-stylesheetEdit.revert.success=Stylesheet erfolgreich zur\u00FCckgesetzt
-stylesheetEdit.revertTip=Falls Sie auf Probleme sto\u00DFen, welche Sie nicht mehr beheben k\u00F6nnen, so beginnen Sie einfach von vorne, indem Sie auf 'Stylesheet wiederherstellen' klicken.
-stylesheetEdit.save.success=Stylesheet erfolgreich aktualisiert
-stylesheetEdit.subtitle=Weblog Stylesheet bearbeiten
-stylesheetEdit.tip=Dieses Formular erlaubt Ihnen das Stylesheet f\u00FCr Ihr Thema zu bearbeiten.
-stylesheetEdit.title=Stylesheet
 tabbedmenu.admin=Administration
 tabbedmenu.admin.cacheInfo=Cache Info
 tabbedmenu.admin.commentManagement=Kommentare
@@ -599,7 +592,6 @@ tabbedmenu.admin.planets=Benutzerdefinierte Gruppen
 tabbedmenu.admin.userAdmin=Benutzerverwaltung
 tabbedmenu.bookmarks.allFolders=Blogroll
 tabbedmenu.design=Design
-tabbedmenu.design.stylesheet=Stylesheet
 tabbedmenu.weblog=Weblog
 tabbedmenu.weblog.archives=Archiv
 tabbedmenu.weblog.categories=Kategorien
