+++ app/src/main/resources/ApplicationResources_es.properties
@@ -549,3 +549,4 @@ yourWebsites.invitationsPrompt=Tiene una o m\u00E1s invitaciones para aceptar o 
 yourWebsites.youAreInvited=Ha sido invitado a unirse al weblog [{0}] -
 yourWebsites.declined=Ha rechazado una invitaci\u00F3n a unirse al weblog [{0}]
 yourWebsites.resigned=Ha renunciado del weblog  [{0}]
+templateEdit.default.success=
