+++ app/src/main/resources/ApplicationResources_ko.properties
@@ -778,16 +778,6 @@ statCount.weblogCommentCountType=\uc6f9\ub85c\uadf8 \uc758\uacac \uac1c\uc218
 statCount.weblogEntryCommentCountType=\uc6f9\ub85c\uadf8 \uae30\uc0ac \uc758\uacac \uc218
 statCount.weblogDayHits=\uae08\uc77c \ubc29\ubb38 \ud68c\uc218
 
-# ------------------------------------------------------------------ Stylesheet Editor
-
-stylesheetEdit.title=\uc2a4\ud0c0\uc77c\uc2dc\ud2b8
-stylesheetEdit.subtitle=\uc6f9\ub85c\uadf8 \ucee4\uc2a4\ud140 \uc2a4\ud0c0\uc77c\uc2dc\ud2b8 \ud3b8\uc9d1
-stylesheetEdit.tip=\uc774 \ud3fc\uc744 \uc0ac\uc6a9\ud558\uc5ec \uadc0\ud558\uc758 \ud14c\ub9c8\ub97c \uc704\ud55c \uc2a4\ud0c0\uc77c\uc2dc\ud2b8\ub97c \ud3b8\uc9d1\ud558\uc2ed\uc2dc\uc624.
-stylesheetEdit.revertTip=\uc5b4\ub5bb\uac8c \ud574\uacb0\ud574\uc57c \ud560\uc9c0 \uc54c \uc218 \uc5c6\ub294 \ubb38\uc81c\uac00 \ubc1c\uc0dd\ud558\uac8c \ub418\uba74, \
-\uc5b8\uc81c\ub4e0\uc9c0 \ubcf5\uad6c \ubc84\ud2bc\uc744 \uc120\ud0dd\ud558\uc5ec \uc7ac\uc2dc\uc791\ud558\uc2dc\uae30 \ubc14\ub78d\ub2c8\ub2e4.
-stylesheetEdit.revert=\uc2a4\ud0c0\uc77c\uc2dc\ud2b8 \ubcf5\uad6c
-stylesheetEdit.save.success=\uc2a4\ud0c0\uc77c\uc2dc\ud2b8\uac00 \uc131\uacf5\uc801\uc73c\ub85c \ubcc0\uacbd\ub418\uc5c8\uc2b5\ub2c8\ub2e4.
-stylesheetEdit.revert.success=\uc2a4\ud0c0\uc77c\uc2dc\ud2b8\uac00 \uc131\uacf5\uc801\uc73c\ub85c \ub418\ub3cc\ub824\uc84c\uc2b5\ub2c8\ub2e4.
 # ------------------------------------------------------------------ Tabbed Menu
 
 tabbedmenu.weblog=\uc0dd\uc131 \ubc0f \ud3b8\uc9d1
@@ -797,7 +787,6 @@ tabbedmenu.weblog.categories=\uce74\ud14c\uace0\ub9ac
 tabbedmenu.bookmarks.allFolders=\ubd81\ub9c8\ud06c
 
 tabbedmenu.design=\ub514\uc790\uc778
-tabbedmenu.design.stylesheet=\uc2a4\ud0c0\uc77c\uc2dc\ud2b8
 
 tabbedmenu.website=\ud658\uacbd \uc124\uc815
 tabbedmenu.website.settings=\ud658\uacbd \uc124\uc815 \uc815\ubcf4
