+++ app/src/main/webapp/WEB-INF/jsps/editor/Templates.jsp
@@ -73,7 +73,7 @@
             <td style="vertical-align:middle"><s:property value="#p.description" /></td>
 
             <td class="center" style="vertical-align:middle">
-                 <s:if test="!#p.required || !customTheme" >
+                 <s:if test="(!#p.required && #p.action.readableName != 'Stylesheet') || !customTheme" >
                      <s:url var="removeUrl" action="templateRemove">
                          <s:param name="weblog" value="actionWeblog.handle"/>
                          <s:param name="removeId" value="#p.id"/>
